+++ src/CrimeWatch.Domain/AggregateModels/ReportAggregate/_using.cs
@@ -1,0 +1,2 @@
+﻿global using CrimeWatch.Domain.AggregateModels.ModeratorAggregate;
+global using CrimeWatch.Domain.AggregateModels.WitnessAggregate;
