--- src/CrimeWatch.Domain/Models/ReportModel/_using.cs
@@ -1,2 +1,0 @@
-﻿global using CrimeWatch.Domain.Models.ModeratorModel;
-global using CrimeWatch.Domain.Models.WitnessModel;
